Add seperate utils file and improve exports
